requirements.txt
django==3.1.6
djangorestframework==3.12.2
django_filter==2.4.0